Include proper version of expected class from precompiled header
